+++ libase/libdsn/dsn_test.go
@@ -1,4 +1,4 @@
-package dsn
+package libdsn
 
 import (
 	"net/url"
