+++ tests/libtest/setup_teardown.go
@@ -5,11 +5,11 @@ import (
 	"database/sql"
 	"fmt"
 
-	"github.com/SAP/go-ase/libase/dsn"
+	"github.com/SAP/go-ase/libase/libdsn"
 )
 
 // SetupDB creates a database and sets .Database on the passed testDsn
-func SetupDB(testDsn *dsn.DsnInfo) error {
+func SetupDB(testDsn *libdsn.DsnInfo) error {
 	db, err := sql.Open("ase", testDsn.AsSimple())
 	if err != nil {
 		return fmt.Errorf("Failed to open database: %v", err)
@@ -45,7 +45,7 @@ func SetupDB(testDsn *dsn.DsnInfo) error {
 
 // TeardownDB deletes the database indicated by .Database of the passed
 // testDsn and unsets the member.
-func TeardownDB(testDsn *dsn.DsnInfo) error {
+func TeardownDB(testDsn *libdsn.DsnInfo) error {
 	db, err := sql.Open("ase", testDsn.AsSimple())
 	if err != nil {
 		return fmt.Errorf("Failed to open database: %v", err)
